requirements.txt
mysql-connector-python
pandas 
openpyxl